1FzGrExv: Changes to comment following review
Andrea and I took a look at the API comments on the areas of code which were changed as part of this task.  We decided to do the following:

- Split create-identity-response-handler out into a file of its own.
- Use the term 'service' rather than 'relying party' (the latter is a SAML term).
- Avoid referring to 'new' services, and explain what we mean when we mention legacy services.
- Make a few other comment changes to improve clarity.

Solo: @alan-gds
