+++ test/create-response-handler-test.ts
@@ -1,5 +1,5 @@
 import * as assert from 'assert'
-import { createResponseHandler, createIdentityResponseHandler, ResponseScenarios, IdentityResponseScenarios } from '../lib/create-response-handler'
+import { createResponseHandler, ResponseScenarios } from '../lib/create-response-handler'
 import { Scenario } from '../lib/verify-service-provider-api/translated-response-body'
 import * as td from 'testdouble'
 
@@ -173,124 +173,6 @@ describe('The createResponseHandler function', () => {
     verifyNotCalled(onCreateUser)
     verifyNotCalled(onAuthnFailed)
     verifyNotCalled(onNoMatch)
-    verifyNotCalled(onCancel)
-  })
-})
-
-describe('The createIdentityResponseHandler function', () => {
-  let scenarios: IdentityResponseScenarios
-  let onIdentityVerified: any
-  let onAuthnFailed: any
-  let onCancel: any
-  let onError: any
-
-  beforeEach(() => {
-    onIdentityVerified = td.function()
-    onAuthnFailed = td.function()
-    onCancel = td.function()
-    onError = td.function()
-
-    scenarios = {
-      onIdentityVerified,
-      onAuthnFailed,
-      onCancel,
-      onError
-    }
-  })
-
-  it('should return a passport authenticate callback function', () => {
-    const result = createIdentityResponseHandler(scenarios)
-    assert.strictEqual(typeof result, 'function')
-    assert.strictEqual(result.length, 4)
-  })
-
-  it('callback should call onMatch when called with an existing user', () => {
-    const error = null as any
-    const identity = {}
-    const info = { scenario: Scenario.IDENTITY_VERIFIED, pid: '', levelOfAssurance: '' }
-    const status = null as any
-
-    createIdentityResponseHandler(scenarios)(error, identity, info, status)
-
-    td.verify(onIdentityVerified(identity))
-
-    verifyNotCalled(onAuthnFailed)
-    verifyNotCalled(onCancel)
-    verifyNotCalled(onError)
-  })
-
-  it('callback should call onAuthnFailed when called with no user and no error', () => {
-    const error = null as any
-    const identity = null
-    const info = Scenario.AUTHENTICATION_FAILED
-    const status = null as any
-
-    createIdentityResponseHandler(scenarios)(error, identity, info, status)
-
-    td.verify(onAuthnFailed())
-
-    verifyNotCalled(onIdentityVerified)
-    verifyNotCalled(onCancel)
-    verifyNotCalled(onError)
-  })
-
-  it('callback should call onCancel when called with no user, no error, and a CANCELLATION scenario', () => {
-    const error = null as any
-    const identity = null
-    const info = Scenario.CANCELLATION
-    const status = null as any
-
-    createIdentityResponseHandler(scenarios)(error, identity, info, status)
-
-    td.verify(onCancel())
-
-    verifyNotCalled(onIdentityVerified)
-    verifyNotCalled(onAuthnFailed)
-    verifyNotCalled(onError)
-  })
-
-  it('callback should call onError when called with no user, no error, and an inappropriate scenario', () => {
-    const error = null as any
-    const identity = null
-    const info = Scenario.SUCCESS_MATCH
-    const status = null as any
-
-    createIdentityResponseHandler(scenarios)(error, identity, info, status)
-
-    td.verify(onError(new Error('Unrecognised Scenario SUCCESS_MATCH')))
-
-    verifyNotCalled(onIdentityVerified)
-    verifyNotCalled(onAuthnFailed)
-    verifyNotCalled(onCancel)
-  })
-
-  it('callback should call onError when called with no user, no error, and a REQUEST_ERROR info', () => {
-    const error = null as any
-    const identity = null
-    const info = Scenario.REQUEST_ERROR
-    const status = null as any
-
-    createIdentityResponseHandler(scenarios)(error, identity, info, status)
-
-    td.verify(onError(new Error('SAML Response was an error')))
-
-    verifyNotCalled(onIdentityVerified)
-    verifyNotCalled(onAuthnFailed)
-    verifyNotCalled(onCancel)
-  })
-
-  it('callback should call onError when called with an error', () => {
-    const error = new Error('some-really-bad-error')
-    const identity = null
-    const info = null as any
-    const status = null as any
-
-    createIdentityResponseHandler(scenarios)(error, identity, info, status)
-
-    td.verify(onError(error))
-
-    verifyNotCalled(onIdentityVerified)
-    verifyNotCalled(onAuthnFailed)
     verifyNotCalled(onCancel)
   })
 })
